Allow to inject sdk method
